# api/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.BusinessType:
    enum:
    - noodle
    - bakery
    - snack
    - drink
    - chinese_food
    type: string
    x-enum-comments:
      BusinessTypeBakery: 烘焙
      BusinessTypeChineseFood: 中餐
      BusinessTypeDrink: 饮品
      BusinessTypeNoodle: 面馆
      BusinessTypeSnack: 小吃
    x-enum-varnames:
    - BusinessTypeNoodle
    - BusinessTypeBakery
    - BusinessTypeSnack
    - BusinessTypeDrink
    - BusinessTypeChineseFood
...
  domain.Store:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/domain.Address'
        description: 地址
      admin_phone_number:
        description: 管理员手机号
        type: string
      business_hours:
        description: 营业时间段
        items:
          $ref: '#/definitions/domain.BusinessHours'
        type: array
      business_license_url:
        description: 营业执照图片地址
        type: string
      business_model:
        allOf:
        - $ref: '#/definitions/domain.BusinessModel'
        description: 经营模式：直营 加盟
      business_type_code:
        allOf:
        - $ref: '#/definitions/domain.BusinessType'
        description: 业态类型
      cashier_desk_url:
        description: 收银台照片地址
        type: string
      contact_name:
        description: 联系人
        type: string
      contact_phone:
        description: 联系电话
        type: string
      created_at:
        type: string
      dining_environment_url:
        description: 就餐环境照片地址
        type: string
      dining_periods:
        description: 就餐时段
        items:
          $ref: '#/definitions/domain.DiningPeriod'
        type: array
      food_operation_license_url:
        description: 食品经营许可证照片地址
        type: string
      id:
        type: string
      location_number:
        description: 门店位置编号
        type: string
      login_account:
        description: 登录账号
        type: string
      login_password:
        description: 登录密码(加密存储)
        type: string
      merchant_id:
        description: 商户 ID
        type: string
      merchant_name:
        description: 商户名称
        type: string
      shift_times:
        description: 班次时间
        items:
          $ref: '#/definitions/domain.ShiftTime'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/domain.StoreStatus'
        description: '状态: 营业 停业'
      store_code:
        description: 门店编码(保留字段)
        type: string
      store_logo:
        description: logo 图片地址
        type: string
      store_name:
        description: 门店名称,长度不超过30个字
        type: string
      store_short_name:
        description: 门店简称
        type: string
      storefront_url:
        description: 门店门头照片地址
        type: string
      unified_social_credit_code:
        description: 统一社会信用代码
        type: string
      updated_at:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: string
      data: {}
    type: object
...
  types.Address:
    properties:
      address:
        description: 详细地址
        maxLength: 100
        type: string
      country:
        allOf:
        - $ref: '#/definitions/domain.Country'
        description: 国家/地区
      lat:
        description: 纬度
        maxLength: 50
        type: string
      lng:
        description: 经度
        maxLength: 50
        type: string
      province:
        allOf:
        - $ref: '#/definitions/domain.Province'
        description: 省份
    type: object
...
  types.CreateStoreReq:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/types.Address'
        description: 地址  创建门店时地址校验必填，创建商户时可选
      admin_phone_number:
        description: 管理员手机号
        maxLength: 20
        type: string
      business_hours:
        description: 营业时间段
        items:
          $ref: '#/definitions/domain.BusinessHours'
        type: array
      business_license_url:
        description: 营业执照图片地址
        maxLength: 500
        type: string
      business_model:
        allOf:
        - $ref: '#/definitions/domain.BusinessModel'
        description: 直营/加盟
      business_type_code:
        allOf:
        - $ref: '#/definitions/domain.BusinessType'
        description: 业务类型
      cashier_desk_url:
        description: 收银台照片地址
        maxLength: 500
        type: string
      contact_name:
        description: 联系人
        maxLength: 255
        type: string
      contact_phone:
        description: 联系电话
        maxLength: 255
        type: string
      dining_environment_url:
        description: 就餐环境照片地址
        maxLength: 500
        type: string
      dining_periods:
        description: 就餐时段
        items:
          $ref: '#/definitions/domain.DiningPeriod'
        type: array
      food_operation_license_url:
        description: 食品经营许可证照片地址
        maxLength: 500
        type: string
      location_number:
        description: 门店位置编号
        maxLength: 255
        type: string
      login_account:
        description: 登录账号
        type: string
      login_password:
        description: 登录密码(加密存储)
        type: string
      shift_times:
        description: 班次时间
        items:
          $ref: '#/definitions/domain.ShiftTime'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/domain.StoreStatus'
        description: 营业/停业
      store_code:
        description: 门店编码(保留字段)
        maxLength: 50
        type: string
      store_logo:
        description: logo 图片地址
        maxLength: 500
        type: string
      store_name:
        description: 门店名称,长度不超过30个字
        maxLength: 30
        type: string
      store_short_name:
        description: 门店简称
        maxLength: 30
        type: string
      storefront_url:
        description: 门店门头照片地址
        maxLength: 500
        type: string
      unified_social_credit_code:
        description: 统一社会信用代码
        maxLength: 255
        type: string
    required:
    - address
    - admin_phone_number
    - business_hours
    - business_model
    - business_type_code
    - dining_periods
    - location_number
    - login_account
    - login_password
    - shift_times
    - status
    - store_name
    type: object
...
  types.MerchantRenewalReq:
    properties:
      merchant_id:
        description: 商户 ID
        type: string
      purchase_duration:
        description: 购买时长
        type: integer
      purchase_duration_unit:
        allOf:
        - $ref: '#/definitions/domain.PurchaseDurationUnit'
        description: 购买时长单位
    required:
    - merchant_id
    - purchase_duration
    - purchase_duration_unit
    type: object
...
  types.UpdateMerchantReq:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/types.Address'
        description: 地址
      admin_phone_number:
        description: 管理员手机号
        maxLength: 20
        type: string
      brand_name:
        description: 品牌名称
        maxLength: 50
        type: string
      business_type_code:
        allOf:
        - $ref: '#/definitions/domain.BusinessType'
        description: 业务类型
      description:
        description: 商户描述(保留字段)
        maxLength: 255
        type: string
      merchant_code:
        description: 商户编号(保留字段)
        maxLength: 50
        type: string
      merchant_logo:
        description: logo 图片地址
        maxLength: 500
        type: string
      merchant_name:
        description: 商户名称,最长不得超过50个字
        maxLength: 50
        type: string
      merchant_short_name:
        description: 商户简称
        maxLength: 50
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.MerchantStatus'
        description: '状态: 正常,停用,过期'
    required:
    - admin_phone_number
    - business_type_code
    - merchant_name
    type: object
  types.UpdateStoreMerchantReq:
    properties:
      merchant:
        $ref: '#/definitions/types.UpdateMerchantReq'
      store:
        $ref: '#/definitions/types.UpdateMStoreReq'
    type: object
  types.UpdateStoreReq:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/types.Address'
        description: 地址
      admin_phone_number:
        description: 管理员手机号
        maxLength: 20
        type: string
      business_hours:
        description: 营业时间段
        items:
          $ref: '#/definitions/domain.BusinessHours'
        type: array
      business_license_url:
        description: 营业执照图片地址
        maxLength: 500
        type: string
      business_model:
        allOf:
        - $ref: '#/definitions/domain.BusinessModel'
        description: 直营/加盟
      business_type_code:
        allOf:
        - $ref: '#/definitions/domain.BusinessType'
        description: 业务类型
      cashier_desk_url:
        description: 收银台照片地址
        maxLength: 500
        type: string
      contact_name:
        description: 联系人
        maxLength: 255
        type: string
      contact_phone:
        description: 联系电话
        maxLength: 255
        type: string
      dining_environment_url:
        description: 就餐环境照片地址
        maxLength: 500
        type: string
      dining_periods:
        description: 就餐时段
        items:
          $ref: '#/definitions/domain.DiningPeriod'
        type: array
      food_operation_license_url:
        description: 食品经营许可证照片地址
        maxLength: 500
        type: string
      location_number:
        description: 门店位置编号
        maxLength: 255
        type: string
      shift_times:
        description: 班次时间
        items:
          $ref: '#/definitions/domain.ShiftTime'
        type: array
      status:
        allOf:
        - $ref: '#/definitions/domain.StoreStatus'
        description: 营业/停业
      store_code:
        description: 门店编码(保留字段)
        maxLength: 50
        type: string
      store_logo:
        description: logo 图片地址
        maxLength: 500
        type: string
      store_name:
        description: 门店名称,长度不超过30个字
        maxLength: 30
        type: string
      store_short_name:
        description: 门店简称
        maxLength: 30
        type: string
      storefront_url:
        description: 门店门头照片地址
        maxLength: 500
        type: string
      unified_social_credit_code:
        description: 统一社会信用代码
        maxLength: 255
        type: string
    required:
    - address
    - admin_phone_number
    - business_hours
    - business_model
    - business_type_code
    - dining_periods
    - location_number
    - shift_times
    - status
    - store_name
    type: object
info:
  contact: {}
  description: 供品牌商后台调用.
  title: 品牌商后台 API
  version: "1.0"
paths:
...
  /merchant/brand:
    put:
      consumes:
      - application/json
      description: 更新品牌商户
      parameters:
      - description: 商户ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新品牌商户请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateMerchantReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 更新品牌商户
      tags:
      - 商户管理
...
  /merchant/renewal:
    post:
      consumes:
      - application/json
      description: 商户续费续期
      parameters:
      - description: 商户续期请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.MerchantRenewalReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 商户续期
      tags:
      - 商户管理
  /merchant/store:
    post:
      consumes:
      - application/json
      description: 创建单个门店
      parameters:
      - description: 创建门店请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.CreateStoreReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 创建门店
      tags:
      - 门店管理
    put:
      consumes:
      - application/json
      description: 更新门店商户（商户 + 门店）
      parameters:
      - description: 商户ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新门店商户请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateStoreMerchantReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 更新门店商户
      tags:
      - 商户管理
  /merchant/store/{id}:
    delete:
      consumes:
      - application/json
      description: 删除单个门店
      parameters:
      - description: 门店ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: No Content
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: 删除门店
      tags:
      - 门店管理
    get:
      consumes:
      - application/json
      description: 根据门店ID获取门店信息
      parameters:
      - description: 门店ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.Store'
              type: object
      security:
      - BearerAuth: []
      summary: 获取门店
      tags:
      - 门店管理
    put:
      consumes:
      - application/json
      description: 更新单个门店
      parameters:
      - description: 门店ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新门店请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.UpdateStoreReq'
      produces:
      - application/json
      responses:
        "200":
          description: No Content
      security:
      - BearerAuth: []
      summary: 更新门店
      tags:
      - 门店管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"